# openapi.yaml
openapi: 3.1.0
info:
  contact:
    email: dev@example.com
    name: Development Team
  description: RESTful API for billing service
  title: Billing Service API
  version: "1.0"
servers:
  - url: http://localhost:8080
    description: Generated server url
...
paths:
...
  /api/subscriptions:
    post:
      description: Subscribes a customer to a specific product, optionally linked to a device
      operationId: createSubscription
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SubscriptionRequest'
        required: true
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/SubscriptionResponse'
          description: Subscription created successfully
        "400":
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ValidationErrorResponse'
          description: Invalid request data
        "404":
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Customer, Product, or Device not found
      summary: Create a subscription
      tags:
        - Subscription Management
  /api/subscriptions/customer/{customerId}:
    get:
      description: Retrieves all subscriptions for a specific customer
      operationId: getCustomerSubscriptions
      parameters:
        - in: path
          name: customerId
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SubscriptionResponse'
          description: Successfully retrieved subscriptions
        "404":
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Customer not found
      summary: List customer subscriptions
      tags:
        - Subscription Management
  /api/subscriptions/{id}:
    get:
      description: Retrieves details of a specific subscription
      operationId: getSubscription
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/SubscriptionResponse'
          description: Subscription found
        "404":
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Subscription not found
      summary: Get subscription by ID
      tags:
        - Subscription Management
components:
  schemas:
...
    ErrorResponse:
      type: object
      description: Response model for general API errors (e.g., 404, 500)
      properties:
        message:
          type: string
          description: Error message
          example: Resource not found
        path:
          type: string
          description: Path where the error occurred
          example: /api/customers/999
        status:
          type: integer
          format: int32
          description: HTTP Status code
          example: 404
        timestamp:
          type: string
          format: date-time
          description: Timestamp of the error
...
    SubscriptionRequest:
      type: object
      description: Request model for creating a new subscription
      properties:
        baseFee:
          type: number
          description: Base fee at time of subscription
          example: "29.99"
          minimum: 0.0
        customerId:
          type: integer
          format: int64
          description: ID of the subscribing customer
          example: 1
        deviceId:
          type: integer
          format: int64
          description: Optional ID of the primary device for this subscription
          example: 1
        discountRate:
          type: number
          description: Applied discount rate
          example: "0.90"
          minimum: 0.0
        endDate:
          type: string
          format: date-time
          description: When the subscription expires
          example: "2027-03-16T00:00:00Z"
        periodDays:
          type: integer
          format: int32
          description: Duration of the subscription in days
          example: 365
          minimum: 1
        productId:
          type: integer
          format: int64
          description: ID of the product to subscribe to
          example: 1
        startDate:
          type: string
          format: date-time
          description: When the subscription begins
          example: "2026-03-16T00:00:00Z"
        status:
          type: string
          description: Current state of the subscription
          enum:
            - PENDING
            - ACTIVE
            - CANCELLED
            - EXPIRED
          example: ACTIVE
      required:
        - customerId
        - periodDays
        - productId
        - startDate
    SubscriptionResponse:
      type: object
      description: Response model representing a customer's subscription
      properties:
        baseFee:
          type: number
          description: Base fee at subscription time
          example: "29.99"
        customerId:
          type: integer
          format: int64
          description: ID of the customer
          example: 1
        deviceId:
          type: integer
          format: int64
          description: ID of the primary device (if any)
          example: 1
        discountRate:
          type: number
          description: Locked-in discount rate
          example: "0.90"
        endDate:
          type: string
          format: date-time
          description: Expiry date
        id:
          type: integer
          format: int64
          description: Unique database identifier for the subscription
          example: 1
        periodDays:
          type: integer
          format: int32
          description: Subscription period in days
          example: 365
        productId:
          type: integer
          format: int64
          description: ID of the subscribed product
          example: 1
        startDate:
          type: string
          format: date-time
          description: Activation date
        status:
          type: string
          description: Current lifecycle status of the subscription
          enum:
            - PENDING
            - ACTIVE
            - CANCELLED
            - EXPIRED
          example: ACTIVE
        totalFee:
          type: number
          description: Total fee (baseFee * discountRate)
          example: "26.99"
    ValidationErrorResponse:
      type: object
      description: Response model for validation errors (400 Bad Request)
      properties:
        errors:
          type: object
          additionalProperties:
            type: string
          description: Detailed field-level validation errors
        message:
          type: string
          description: Overall error message
          example: Validation failed
        path:
          type: string
          description: Path where the error occurred
          example: /api/customers
        status:
          type: integer
          format: int32
          description: HTTP Status code
          example: 400
        timestamp:
          type: string
          format: date-time
          description: Timestamp of the error